+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ## Bug Fixes
 
 ## Enhancements
+* Remove corrupt characters before processing the XSLT #980. Runtime parameter needs to be set to true:-
+      bordertech.wcomponents.xslt.allow.corrupt.characters
 
 # Release 1.2.13